Dont update currentCenter and Zoom in updateMapBounds function
